CPR-1167 refactor
