+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/cprdomainevents/listeners/AddressEventPublisher.kt
@@ -10,9 +10,11 @@ import uk.gov.justice.digital.hmpps.personrecord.extensions.UK_ZONE
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.AddressEntity
 import uk.gov.justice.digital.hmpps.personrecord.model.types.SourceSystemType
 import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.address.AddressCreated
+import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.address.AddressDeleted
 import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.address.AddressUpdated
 import uk.gov.justice.digital.hmpps.personrecord.service.queue.DomainEventPublisher
 import uk.gov.justice.digital.hmpps.personrecord.service.type.CPR_PROBATION_ADDRESS_CREATED
+import uk.gov.justice.digital.hmpps.personrecord.service.type.CPR_PROBATION_ADDRESS_DELETED
 import uk.gov.justice.digital.hmpps.personrecord.service.type.CPR_PROBATION_ADDRESS_UPDATED
 import java.time.Instant
 import java.time.format.DateTimeFormatter
@@ -21,6 +23,7 @@ interface AddressEventPublisher {
   val sourceSystemType: SourceSystemType
   fun onCreate(addressCreated: AddressCreated)
   fun onUpdate(addressUpdated: AddressUpdated)
+  fun onDelete(addressDeleted: AddressDeleted)
 }
 
 @Component
@@ -31,7 +34,7 @@ class ProbationAddressEventPublisher(
   override val sourceSystemType = SourceSystemType.DELIUS
 
   override fun onCreate(addressCreated: AddressCreated) {
-    publishAddressDomainEvent(
+    publishAddressUpsertDomainEvent(
       addressEntity = addressCreated.addressEntity,
       eventSource = addressCreated.eventSource.identifier,
       action = "created",
@@ -40,7 +43,7 @@ class ProbationAddressEventPublisher(
   }
 
   override fun onUpdate(addressUpdated: AddressUpdated) {
-    publishAddressDomainEvent(
+    publishAddressUpsertDomainEvent(
       addressEntity = addressUpdated.addressEntity,
       eventSource = addressUpdated.eventSource.identifier,
       action = "updated",
@@ -48,7 +51,15 @@ class ProbationAddressEventPublisher(
     )
   }
 
-  private fun publishAddressDomainEvent(
+  override fun onDelete(addressDeleted: AddressDeleted) {
+    publishAddressDeleteDomainEvent(
+      addressEntity = addressDeleted.addressEntity,
+      sourceSystemId = addressDeleted.personEntity.extractSourceSystemId()!!,
+      eventSource = addressDeleted.eventSource.identifier,
+    )
+  }
+
+  private fun publishAddressUpsertDomainEvent(
     addressEntity: AddressEntity,
     eventSource: String,
     action: String,
@@ -74,6 +85,33 @@ class ProbationAddressEventPublisher(
           ),
         ),
       ),
+      attributes = mapOf("eventSource" to eventSource),
+    )
+  }
+
+  private fun publishAddressDeleteDomainEvent(
+    addressEntity: AddressEntity,
+    sourceSystemId: String,
+    eventSource: String,
+  ) {
+    val addressId = addressEntity.updateId
+
+    val domainEvent = DomainEvent(
+      eventType = CPR_PROBATION_ADDRESS_DELETED,
+      description = "A probation address has been deleted for a person",
+      occurredAt = DateTimeFormatter.ISO_OFFSET_DATE_TIME.withZone(UK_ZONE).format(Instant.now()),
+      additionalInformation = AdditionalInformation(
+        outboundCprAddressId = addressId.toString(),
+        outboundDeliusAddressId = addressEntity.deliusAddressId,
+      ),
+      personReference = PersonReference(
+        identifiers = listOf(
+          PersonIdentifier("CRN", sourceSystemId),
+        ),
+      ),
+    )
+    domainEventPublisher.publish(
+      domainEvent = domainEvent,
       attributes = mapOf("eventSource" to eventSource),
     )
   }
